// apps/frontend/src/api/models/scrapedJobDTO.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * FastAPI
 * OpenAPI spec version: 0.1.0
 */
import type { ScrapedJobDTODescription } from './scrapedJobDTODescription';
import type { ScrapedJobDTOSource } from './scrapedJobDTOSource';
import type { ScrapedJobDTOPostedAt } from './scrapedJobDTOPostedAt';

export interface ScrapedJobDTO {
  title: string;
  application_url: string;
  company: string;
  description?: ScrapedJobDTODescription;
  source?: ScrapedJobDTOSource;
  posted_at?: ScrapedJobDTOPostedAt;
}